+++ sources/gameloop/gameloop.c
@@ -6,7 +6,7 @@
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/30 01:01:19 by lray              #+#    #+#             */
-/*   Updated: 2024/02/13 13:17:34 by lray             ###   ########.fr       */
+/*   Updated: 2024/02/13 23:36:43 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,6 @@ int	gameloop(t_ctx *ctx)
 	}
 	if (put_img_to_img(frame, ctx->textures[0], 0, 0) != 0)
 	{
-		ft_putstr_fd("Error when adding pixels to frame\n", 2);
 		ctx_free(ctx);
 		exit(EXIT_FAILURE);
 	}
